patch 8.1.2233: cannot get the Vim command line arguments
Problem:    Cannot get the Vim command line arguments.
Solution:   Add v:argv. (Dmitri Vereshchagin, closes #1322)

+++ src/normal.c
@@ -5986,7 +5986,7 @@ nv_g_cmd(cmdarg_T *cap)
 	    oap->motion_type = MCHAR;
 	    oap->inclusive = FALSE;
 	    if (has_mbyte)
-		i = mb_string2cells(ptr, STRLEN(ptr));
+		i = mb_string2cells(ptr, (int)STRLEN(ptr));
 	    else
 		i = (int)STRLEN(ptr);
 	    if (cap->count0 > 0 && cap->count0 <= 100)
